+++ R/shortcut.R
@@ -8,12 +8,12 @@
 #' intermediate [quotation] objects.
 #'
 #' Throughout this package, some functions come in two forms, a "bare"
-#' version which quote their first argument literally, and a
+#' version which quotes its first argument literally, and a
 #' normally-evaluating version with a trailing underscore in its
 #' name. So `is_forced(x)` chiecks whether "x" is a missing variable,
-#' while `is_forced_(x, environment())` checks whether "x"
-#' contains the _name_ of another variable which is missing. The
-#' following are all equivalent:
+#' while `is_forced_(x, environment())` checks whether "x" contains
+#' the _name_ of another variable which is missing. The following are
+#' all equivalent:
 #'
 #' * `arg_env(x)`
 #' * `{y <- quo(x); arg_env_(y)}`